+++ substratevm/src/com.oracle.svm.core.genscavenge/src/com/oracle/svm/core/genscavenge/Space.java
@@ -61,7 +61,7 @@ import jdk.graal.compiler.word.Word;
 public final class Space {
     private final String name;
     private final String shortName;
-    private final boolean isFromSpace;
+    private final boolean isToSpace;
     private final int age;
     private final ChunksAccounting accounting;
 
@@ -77,16 +77,16 @@ public final class Space {
      * collections so they should not move.
      */
     @Platforms(Platform.HOSTED_ONLY.class)
-    Space(String name, String shortName, boolean isFromSpace, int age) {
-        this(name, shortName, isFromSpace, age, null);
+    Space(String name, String shortName, boolean isToSpace, int age) {
+        this(name, shortName, isToSpace, age, null);
     }
 
     @Platforms(Platform.HOSTED_ONLY.class)
-    Space(String name, String shortName, boolean isFromSpace, int age, ChunksAccounting parentAccounting) {
+    Space(String name, String shortName, boolean isToSpace, int age, ChunksAccounting parentAccounting) {
         assert name != null : "Space name should not be null.";
         this.name = name;
         this.shortName = shortName;
-        this.isFromSpace = isFromSpace;
+        this.isToSpace = isToSpace;
         this.age = age;
         this.accounting = new ChunksAccounting(parentAccounting);
     }
@@ -132,6 +132,12 @@ public final class Space {
         return age == (HeapParameters.getMaxSurvivorSpaces() + 1);
     }
 
+    @AlwaysInline("GC performance.")
+    @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+    public boolean isCompactingOldSpace() {
+        return SerialGCOptions.useCompactingOldGen() && isOldSpace();
+    }
+
     @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
     int getAge() {
         return age;
@@ -142,9 +148,15 @@ public final class Space {
         return age + 1;
     }
 
+    @AlwaysInline("GC performance.")
     @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
     boolean isFromSpace() {
-        return isFromSpace;
+        return !isToSpace && !isCompactingOldSpace();
+    }
+
+    @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+    boolean isToSpace() {
+        return isToSpace;
     }
 
     public boolean walkObjects(ObjectVisitor visitor) {
@@ -165,6 +177,17 @@ public final class Space {
         return true;
     }
 
+    boolean walkAlignedHeapChunks(AlignedHeapChunk.Visitor visitor) {
+        AlignedHeapChunk.AlignedHeader chunk = getFirstAlignedHeapChunk();
+        while (chunk.isNonNull()) {
+            if (!visitor.visitChunk(chunk)) {
+                return false;
+            }
+            chunk = HeapChunk.getNext(chunk);
+        }
+        return true;
+    }
+
     public void logUsage(Log log, boolean logIfEmpty) {
         UnsignedWord chunkBytes;
         if (isEdenSpace() && !VMOperation.isGCInProgress()) {
@@ -181,8 +204,8 @@ public final class Space {
     }
 
     public void logChunks(Log log) {
-        HeapChunkLogging.logChunks(log, getFirstAlignedHeapChunk(), shortName, isFromSpace);
-        HeapChunkLogging.logChunks(log, getFirstUnalignedHeapChunk(), shortName, isFromSpace);
+        HeapChunkLogging.logChunks(log, getFirstAlignedHeapChunk(), shortName, isToSpace());
+        HeapChunkLogging.logChunks(log, getFirstUnalignedHeapChunk(), shortName, isToSpace());
     }
 
     /**
@@ -373,9 +396,9 @@ public final class Space {
     /** Promote an aligned Object to this Space. */
     @AlwaysInline("GC performance")
     @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
-    Object promoteAlignedObject(Object original, Space originalSpace) {
+    Object copyAlignedObject(Object original, Space originalSpace) {
         assert ObjectHeaderImpl.isAlignedObject(original);
-        assert this != originalSpace && originalSpace.isFromSpace();
+        assert originalSpace.isFromSpace() || (originalSpace == this && isCompactingOldSpace());
 
         Object copy = copyAlignedObject(original);
         if (copy != null) {
@@ -423,10 +446,19 @@ public final class Space {
             ObjectHeaderImpl.getObjectHeaderImpl().setIdentityHashInField(copy);
         }
         if (isOldSpace()) {
-            // If the object was promoted to the old gen, we need to take care of the remembered
-            // set bit and the first object table (even when promoting from old to old).
-            AlignedHeapChunk.AlignedHeader copyChunk = AlignedHeapChunk.getEnclosingChunk(copy);
-            RememberedSet.get().enableRememberedSetForObject(copyChunk, copy);
+            if (SerialGCOptions.useCompactingOldGen() && GCImpl.getGCImpl().isCompleteCollection()) {
+                /*
+                 * In a compacting complete collection, the remembered set bit is set already during
+                 * marking and the first object table is built later during compaction.
+                 */
+            } else {
+                /*
+                 * If an object was copied to the old generation, its remembered set bit must be set
+                 * and the first object table must be updated (even when copying from old to old).
+                 */
+                AlignedHeapChunk.AlignedHeader copyChunk = AlignedHeapChunk.getEnclosingChunk(copy);
+                RememberedSet.get().enableRememberedSetForObject(copyChunk, copy);
+            }
         }
         return copy;
     }
@@ -502,6 +534,7 @@ public final class Space {
             appendUnalignedHeapChunk(uChunk);
             uChunk = next;
         }
+        assert src.isEmpty();
     }
 
     /**
@@ -562,5 +595,26 @@ public final class Space {
             uChunk = HeapChunk.getNext(uChunk);
         }
         return result;
+    }
+
+    boolean contains(Pointer p) {
+        AlignedHeapChunk.AlignedHeader aChunk = getFirstAlignedHeapChunk();
+        while (aChunk.isNonNull()) {
+            Pointer start = AlignedHeapChunk.getObjectsStart(aChunk);
+            if (start.belowOrEqual(p) && p.belowThan(HeapChunk.getTopPointer(aChunk))) {
+                return true;
+            }
+            aChunk = HeapChunk.getNext(aChunk);
+        }
+
+        UnalignedHeapChunk.UnalignedHeader uChunk = getFirstUnalignedHeapChunk();
+        while (uChunk.isNonNull()) {
+            Pointer start = UnalignedHeapChunk.getObjectStart(uChunk);
+            if (start.belowOrEqual(p) && p.belowThan(HeapChunk.getTopPointer(uChunk))) {
+                return true;
+            }
+            uChunk = HeapChunk.getNext(uChunk);
+        }
+        return false;
     }
 }