+++ substratevm/src/com.oracle.svm.truffle.nfi/src/com/oracle/svm/truffle/nfi/NativeClosure.java
@@ -95,7 +95,9 @@ final class NativeClosure {
 
     private final ClosureArgType[] argTypes;
 
-    private NativeClosure(CallTarget callTarget, Object receiver, ClosureArgType[] argTypes) {
+    private final Target_com_oracle_truffle_nfi_backend_libffi_LibFFILanguage language;
+
+    private NativeClosure(CallTarget callTarget, Object receiver, ClosureArgType[] argTypes, Target_com_oracle_truffle_nfi_backend_libffi_LibFFILanguage language) {
         this.callTarget = new WeakReference<>(callTarget);
         if (receiver != null) {
             this.receiver = new WeakReference<>(receiver);
@@ -103,6 +105,7 @@ final class NativeClosure {
             this.receiver = null;
         }
         this.argTypes = argTypes;
+        this.language = language;
     }
 
     /**
@@ -130,7 +133,7 @@ final class NativeClosure {
             }
         }
 
-        NativeClosure closure = new NativeClosure(callTarget, receiver, argTypes);
+        NativeClosure closure = new NativeClosure(callTarget, receiver, argTypes, ctx.language);
         NativeClosureHandle handle = ImageSingletons.lookup(TruffleNFISupport.class).createClosureHandle(closure);
 
         WordPointer codePtr = UnsafeStackValue.get(WordPointer.class);
@@ -276,24 +279,33 @@ final class NativeClosure {
 
     private static void doInvokeClosureBufferRet(Pointer ret, WordPointer args, ClosureData user) {
         NativeClosure closure = lookup(user);
-        Target_com_oracle_truffle_nfi_backend_libffi_NativeArgumentBuffer_Pointer retBuffer = new Target_com_oracle_truffle_nfi_backend_libffi_NativeArgumentBuffer_Pointer(ret.rawValue());
-        Target_com_oracle_truffle_nfi_backend_libffi_LibFFIClosure_RetPatches patches = (Target_com_oracle_truffle_nfi_backend_libffi_LibFFIClosure_RetPatches) closure.call(args, retBuffer);
+        try {
+            Target_com_oracle_truffle_nfi_backend_libffi_NativeArgumentBuffer_Pointer retBuffer = new Target_com_oracle_truffle_nfi_backend_libffi_NativeArgumentBuffer_Pointer(ret.rawValue());
+            Target_com_oracle_truffle_nfi_backend_libffi_LibFFIClosure_RetPatches patches = (Target_com_oracle_truffle_nfi_backend_libffi_LibFFIClosure_RetPatches) closure.call(args, retBuffer);
 
-        if (patches != null) {
-            for (int i = 0; i < patches.count; i++) {
-                Target_com_oracle_truffle_nfi_backend_libffi_NativeArgumentBuffer_TypeTag tag = getTag(patches.patches[i]);
-                int offset = getOffset(patches.patches[i]);
-                Object obj = patches.objects[i];
+            if (patches != null) {
+                for (int i = 0; i < patches.count; i++) {
+                    Target_com_oracle_truffle_nfi_backend_libffi_NativeArgumentBuffer_TypeTag tag = getTag(patches.patches[i]);
+                    int offset = getOffset(patches.patches[i]);
+                    Object obj = patches.objects[i];
 
-                if (tag == Target_com_oracle_truffle_nfi_backend_libffi_NativeArgumentBuffer_TypeTag.OBJECT) {
-                    WordBase handle = ImageSingletons.lookup(TruffleNFISupport.class).createGlobalHandle(obj);
-                    ret.writeWord(offset, handle);
-                } else if (tag == Target_com_oracle_truffle_nfi_backend_libffi_NativeArgumentBuffer_TypeTag.STRING) {
-                    ret.writeWord(offset, serializeStringRet(obj));
-                } else {
-                    // nothing to do
+                    if (tag == Target_com_oracle_truffle_nfi_backend_libffi_NativeArgumentBuffer_TypeTag.OBJECT) {
+                        WordBase handle = ImageSingletons.lookup(TruffleNFISupport.class).createGlobalHandle(obj);
+                        ret.writeWord(offset, handle);
+                    } else if (tag == Target_com_oracle_truffle_nfi_backend_libffi_NativeArgumentBuffer_TypeTag.STRING) {
+                        ret.writeWord(offset, serializeStringRet(obj));
+                    } else {
+                        // nothing to do
+                    }
                 }
             }
+        } catch (Throwable t) {
+            /*
+             * Normally no exceptions can happen here, since the NFI frontend already does all the
+             * exception handling. But certain exceptions can slip through to here, e.g. stack
+             * overflows or async exceptions from Truffle safepoints.
+             */
+            closure.language.getNFIState().setPendingException(t);
         }
     }
 
@@ -340,7 +352,17 @@ final class NativeClosure {
     }
 
     private static void doInvokeClosureVoidRet(WordPointer args, ClosureData user) {
-        lookup(user).call(args, null);
+        NativeClosure closure = lookup(user);
+        try {
+            closure.call(args, null);
+        } catch (Throwable t) {
+            /*
+             * Normally no exceptions can happen here, since the NFI frontend already does all the
+             * exception handling. But certain exceptions can slip through to here, e.g. stack
+             * overflows or async exceptions from Truffle safepoints.
+             */
+            closure.language.getNFIState().setPendingException(t);
+        }
     }
 
     @CEntryPoint(include = CEntryPoint.NotIncludedAutomatically.class, publishAs = Publish.NotPublished)
@@ -386,8 +408,18 @@ final class NativeClosure {
     }
 
     private static void doInvokeClosureStringRet(WordPointer ret, WordPointer args, ClosureData user) {
-        Object retValue = lookup(user).call(args, null);
-        ret.write(serializeStringRet(retValue));
+        NativeClosure closure = lookup(user);
+        try {
+            Object retValue = closure.call(args, null);
+            ret.write(serializeStringRet(retValue));
+        } catch (Throwable t) {
+            /*
+             * Normally no exceptions can happen here, since the NFI frontend already does all the
+             * exception handling. But certain exceptions can slip through to here, e.g. stack
+             * overflows or async exceptions from Truffle safepoints.
+             */
+            closure.language.getNFIState().setPendingException(t);
+        }
     }
 
     @CEntryPoint(include = CEntryPoint.NotIncludedAutomatically.class, publishAs = Publish.NotPublished)
@@ -433,12 +465,22 @@ final class NativeClosure {
     }
 
     private static void doInvokeClosureObjectRet(WordPointer ret, WordPointer args, ClosureData user) {
-        Object obj = lookup(user).call(args, null);
-        if (obj == null) {
-            ret.write(WordFactory.zero());
-        } else {
-            TruffleObjectHandle handle = ImageSingletons.lookup(TruffleNFISupport.class).createGlobalHandle(obj);
-            ret.write(handle);
+        NativeClosure closure = lookup(user);
+        try {
+            Object obj = closure.call(args, null);
+            if (obj == null) {
+                ret.write(WordFactory.zero());
+            } else {
+                TruffleObjectHandle handle = ImageSingletons.lookup(TruffleNFISupport.class).createGlobalHandle(obj);
+                ret.write(handle);
+            }
+        } catch (Throwable t) {
+            /*
+             * Normally no exceptions can happen here, since the NFI frontend already does all the
+             * exception handling. But certain exceptions can slip through to here, e.g. stack
+             * overflows or async exceptions from Truffle safepoints.
+             */
+            closure.language.getNFIState().setPendingException(t);
         }
     }